backend/requirements.txt
aniso8601==7.0.0
backcall==0.1.0
boto3==1.13.10
botocore==1.16.10
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Django==2.2.6
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-extensions==2.2.8
django-filter==2.2.0
django-graphql-jwt==0.3.0
django-storages==1.9.1
docutils==0.15.2
graphene==2.1.8
graphene-django==2.9.0
graphene-file-upload==1.2.2
graphql-core==2.3.1
graphql-relay==2.0.1
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
jmespath==0.10.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
promise==2.3
prompt-toolkit==3.0.4
psycopg2-binary==2.8.4
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python-resize-image==1.1.19
pytz==2019.3
requests==2.23.0
Rx==1.6.1
s3transfer==0.3.3
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.9
wcwidth==0.1.8
gunicorn==20.0.4
